improve extract_overlays.py
- split code into submodules
- externalize overlays into a yaml file to make future expansion easier
- use argparse to implement command-line parsing
- leverage mmap to avoid writing the compressed overlays to disk
- leverage with..as to autoclose resources
