+++ tools/scripts/overlays.yaml
@@ -1,0 +1,22 @@
+---
+overlays:
+  compressed:
+    - name: BATTLING.BIN
+      sector_id: 0x0D
+    - name: FIELD.BIN
+      sector_id: 0x0E
+    - name: WORLD_MAP.BIN
+      sector_id: 0x0F
+    - name: BATTLE.BIN
+      sector_id: 0x10
+    - name: MENU.BIN
+      sector_id: 0x11
+    - name: MOVIE.BIN
+      sector_id: 0x12
+  uncompressed:
+    - name: member_change_menu.bin
+      sector_id: 0x06
+      index: 0x10
+    - name: shop_menu.bin
+      index: 0x10
+      sector_id: 0x09
